+++ SplitDiff/SplitDiff/app-example/(tabs)/explore.tsx
